requirements.txt
Flask[async]
htmx
pydobot
TinyDB
pyserial